Implemented Nick's recommendations for edits as described here from his email
It might be nice to include some install instructions for acquiring the required dependencies (python3, pandas, numpy, matplotlib, etc)

I’m getting an error when running the script out of the box due to the OutputData directory not existing. You could include the OutputData directory in the repo with a readme (so it doesn’t get ignored by git).

Right now the file paths for writing the results are hard coded unix paths and so would fail if running on a Windows OS. You could look into using the os.path.join() method or the `pathlib.Path()` object to make the path OS independent.

Might make sense to give an example of how to specify the number of days/start date from this line in the README: “Defaults to plotting the first week of data or alternatively the user can specify the number of days and the start date”